Refactor move the rest of the site utilities components to the component directory (#104)
# Pull Request

## 📖 Description

<!--- Provide some background and a description of your work. -->
This change moves the rest of the components in the site-utilities folder into the components directory.

The components:
- Dimension
- Direction Switch
- Logo
- Theme selector
- Transparency toggle

All components can be found in the top bar, except the logo which is in the top of the left rail.

## 👩‍💻 Reviewer Notes

<!---
Provide some notes for reviewers to help them provide targeted feedback and testing.
-->
Check the generated site for this PR to see if all components are loading correctly.

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [ ] I have added tests for my changes.
- [x] I have tested my changes.
- [ ] I have updated the project documentation to reflect my changes.
