+++ app/components/dimension/dimension.props.ts
