+++ app/components/direction-switch/direction-switch.props.ts
