+++ app/components/logo/logo.props.ts
