+++ app/components/transparency-toggle/transparency-toggle.props.ts
