--- app/site-utilities/components/index.ts
@@ -1,5 +1,0 @@
-export * from "./dimension";
-export * from "./direction-switch";
-export * from "./logo";
-export * from "./theme-selector";
-export * from "./transparency-toggle";
